backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "passlib[argon2]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
